Fixed incorrect appendix (I-P) references Validated Exploitable Data Flow (VXDF) Format.md
In the "VXDF Document Structure" section, under "3. Supporting Object Definitions," it states: "The formal structures of these objects... are normatively defined in Appendix A: Normative JSON Schema and further described in their respective descriptive appendices (I-P), however the actual descriptive appendices for these controlled vocabularies are Appendices J through Q. 

This seem to be a one-off error that affects even the other appendices (B to I).